Move workspaces from Catalogue Entry to GeoServer Publish Channel.
