+++ README.md
@@ -41,12 +41,6 @@ $ gdal-config --version
 $ pip install gdal==`gdal-config --version`
 ```
 
-#### Setup
-In order for the application to run correctly, at least **1** workspace must
-be created in the Django app which corresponds to the environment variable /
-setting `GEOSERVER_DEFAULT_WORKSPACE_ID`. This must correspond to an existing
-workspace in the application's GeoServer instance.
-
 #### Running
 To run a development server, use the `Django` `manage.py` script:
 ```shell
