+++ govapp/apps/catalogue/serializers/__init__.py
@@ -8,4 +8,3 @@ from . import layer_submissions
 from . import layer_subscriptions
 from . import layer_symbology
 from . import notifications
-from . import workspaces
