+++ govapp/apps/publisher/admin.py
@@ -21,3 +21,4 @@ admin.site.register(models.publish_channels.CDDPPublishChannel, reversion.admin.
 admin.site.register(models.publish_channels.GeoServerPublishChannel, reversion.admin.VersionAdmin)
 admin.site.register(models.publish_entries.PublishEntry, PublishEntryAdmin)
 admin.site.register(models.notifications.EmailNotification, reversion.admin.VersionAdmin)
+admin.site.register(models.workspaces.Workspace, reversion.admin.VersionAdmin)
