+++ tests/factories/publisher/__init__.py
@@ -3,3 +3,4 @@
 from . import notifications
 from . import publish_channels
 from . import publish_entries
+from . import workspaces
