+++ tests/factories/publisher/publish_channels.py
@@ -7,13 +7,16 @@ import factory.fuzzy
 
 # Local
 from govapp.apps.publisher import models
+from tests.factories.publisher import workspaces
 
 
 class CDDPPublishChannelFactory(factory.django.DjangoModelFactory):
     """Factory for a CDDP Publish Channel."""
     name = factory.Sequence(lambda n: f"CDDP Publish Channel {n + 1}")
     description = factory.Faker("paragraph")
+    mode = factory.fuzzy.FuzzyChoice(models.publish_channels.CDDPPublishChannelMode)
     frequency = factory.fuzzy.FuzzyChoice(models.publish_channels.PublishChannelFrequency)
+    path = factory.Faker("tests")  # Hardcode for unit tests
 
     class Meta:
         """Publish Channel Factory Metadata."""
@@ -24,7 +27,9 @@ class GeoServerPublishChannelFactory(factory.django.DjangoModelFactory):
     """Factory for a GeoServer Publish Channel."""
     name = factory.Sequence(lambda n: f"GeoServer Publish Channel {n + 1}")
     description = factory.Faker("paragraph")
+    mode = factory.fuzzy.FuzzyChoice(models.publish_channels.GeoServerPublishChannelMode)
     frequency = factory.fuzzy.FuzzyChoice(models.publish_channels.PublishChannelFrequency)
+    workspace = factory.SubFactory(workspaces.WorkspaceFactory)
 
     class Meta:
         """Publish Channel Factory Metadata."""
